requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
psycopg[binary]
pydantic-settings==2.5.2
pdfplumber==0.11.4
python-multipart==0.0.9
python-dotenv==1.0.1
apscheduler==3.10.4
playwright==1.49.1
anthropic>=0.40.0
